requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
autopep8==2.3.2
av==14.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
choreographer==1.0.10
click==8.1.8
clip==1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
cython==3.0.11
cython-bbox==0.1.5
deep-sort-realtime==1.3.2
deep-sort-reid==0.1.5
deepface==0.0.93
et-xmlfile==2.0.0
faiss-cpu==1.11.0.post1
filelock==3.17.0
fire==0.7.0
flake8==7.1.1
flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iopath==0.1.10
isort==4.3.21
itsdangerous==2.2.0
jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.0
kaleido==1.0.0
keras==3.9.2
kiwisolver==1.4.8
lap==0.5.12
libclang==18.1.1
logistro==1.1.0
loss==0.1.2
lz4==4.4.4
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.5.1
motmetrics==1.4.0
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.9
narwhals==2.2.0
networkx==3.4.2
numpy==2.1.1
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.9.79
onnx==1.17.0
onnxruntime-gpu==1.21.1
onnxslim==0.1.59
opencv-python==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.4.0
optree==0.15.0
orjson==3.11.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==6.3.0
portalocker==3.2.0
progress==1.6.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-core==2.27.2
pyflakes==3.2.0
pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
pywin32==311
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.1
screeninfo==0.8.1
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
soupsieve==2.6
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.19.0a20250209
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.4
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorrt==10.12.0.36
tensorrt-cu11==10.12.0.36
tensorrt-cu11-bindings==10.12.0.36
tensorrt-cu11-libs==10.12.0.36
tensorrt-cu12==10.12.0.36
tensorrt-cu12-bindings==10.12.0.36
tensorrt-cu12-libs==10.12.0.36
termcolor==3.1.0
tf-keras==2.19.0
threadpoolctl==3.6.0
timm==1.0.17
tomli==2.2.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchreid==0.2.5
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
yacs==0.1.8
yapf==0.43.0
moondream
